LayoutTests/http/tests/referrer-policy-anchor/no-referrer-when-downgrade/cross-origin-http.https-expected.txt

Tests the behavior of no-referrer-when-downgrade referrer policy.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS referrer is expected
PASS successfullyParsed is true

TEST COMPLETE
